--- README.md
@@ -1,4 +1,0 @@
-# Basecode Repository
-This repository contains some basic code for Hardware Class (with drive motors), a basic teleop, PID stuff, and some Auton.
-
-Feel free to use this as a template for your FTC Repository! :D
