+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/subsystems/vision/SleeveDetection.java
@@ -1,4 +1,4 @@
-package org.firstinspires.ftc.teamcode.auton;
+package org.firstinspires.ftc.teamcode.subsystems.vision;
 
 import org.opencv.core.Core;
 import org.opencv.core.Mat;
